# DOD7/backend/swagger.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "Swagger GIP 2022",
		"version": "1.0.11"
	},
	"servers": [
		{
			"url": "http://localhost:5000/api-docs"
		}
	],
	"paths": {
		"/document": {
			"put": {
				"tags": [
					"document"
				],
				"summary": "Update an existing document",
				"description": "Update an existing document by Id",
				"operationId": "updatedocument",
				"requestBody": {
					"description": "Update an existent document in the position",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/document"
							}
						},
						"application/xml": {
							"schema": {
								"$ref": "#/components/schemas/document"
							}
						},
						"application/x-www-form-urlencoded": {
							"schema": {
								"$ref": "#/components/schemas/document"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Successful operation",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/document"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/document"
								}
							}
						}
					},
					"400": {
						"description": "Invalid ID supplied"
					},
					"404": {
						"description": "document not found"
					},
					"405": {
						"description": "Validation exception"
					}
				}
			},
			"post": {
				"tags": [
					"document"
				],
				"summary": "Add a new document to the position",
				"description": "Add a new document to the position",
				"operationId": "adddocument",
				"requestBody": {
					"description": "Create a new document in the position",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/document"
							}
						},
						"application/xml": {
							"schema": {
								"$ref": "#/components/schemas/document"
							}
						},
						"application/x-www-form-urlencoded": {
							"schema": {
								"$ref": "#/components/schemas/document"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Successful operation",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/document"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/document"
								}
							}
						}
					},
					"405": {
						"description": "Invalid input"
					}
				}
			},
			"get": {
				"tags": [
					"document"
				],
				"summary": "Finds documents by tags",
				"description": "Multiple tags can be provided with comma separated strings. Use tag1, tag2, tag3 for testing.",
				"operationId": "finddocumentsByTags",
				"parameters": [
					{
						"name": "keyword",
						"in": "query",
						"required": false,
						"explode": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "pageSize",
						"in": "query",
						"required": false,
						"explode": true,
						"schema": {
							"type": "number"
						}
					},
					{
						"name": "pageIndex",
						"in": "query",
						"required": false,
						"explode": true,
						"schema": {
							"type": "number"
						}
					}
				],
				"responses": {
					"200": {
						"description": "successful operation",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/document"
									}
								}
							},
							"application/xml": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/document"
									}
								}
							}
						}
					},
					"400": {
						"description": "Invalid tag value"
					}
				}
			}
		},
		"/document/{documentId}": {
			"get": {
				"tags": [
					"document"
				],
				"summary": "Find document by ID",
				"description": "Returns a single document",
				"operationId": "getdocumentById",
				"parameters": [
					{
						"name": "documentId",
						"in": "path",
						"description": "ID of document to return",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					}
				],
				"responses": {
					"200": {
						"description": "successful operation",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/document"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/document"
								}
							}
						}
					},
					"400": {
						"description": "Invalid ID supplied"
					},
					"404": {
						"description": "document not found"
					}
				}
			},
			"delete": {
				"tags": [
					"document"
				],
				"summary": "Deletes a document",
				"description": "delete a document",
				"operationId": "deletedocument",
				"parameters": [
					{
						"name": "documentId",
						"in": "path",
						"description": "document id to delete",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					}
				],
				"responses": {
					"400": {
						"description": "Invalid document value"
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
			"document": {
				"type": "object",
				"properties": {
					"_id": {
						"type": "string",
						"example": "string"
					},
					"name": {
						"type": "string",
						"example": "string"
					},
					"link": {
						"type": "string",
						"example": "string"
					},
					"owner": {
						"type": "string",
						"example": "string"
					},
					"thumbnailLink": {
						"type": "string",
						"example": "string"
					},
					"type": {
						"type": "string",
						"example": "string"
					},
					"createAt": {
						"type": "string",
						"example": "string"
					},
					"updateAt": {
						"type": "string",
						"example": "string"
					}
				},
				"xml": {
					"name": "order"
				}
			},
...
		}
	}
}